apps/backend/pyproject.toml
# pyproject.toml - CompletePay Agent (Apêndice A)
[project]
name = "completepay-agent"
version = "0.1.0"
description = "Agente de IA CompletePay com Agno Framework"
readme = "README.md"
requires-python = ">=3.11"

# Dependencias: agno + LLM (anthropic, openai, google-genai), DB (psycopg, asyncpg, sqlalchemy, pgvector),
# API (fastapi, uvicorn), CLI (typer, rich), config (pydantic, dotenv), guardrails (pyjwt), redis.
dependencies = [
    "agno>=2.5.0,<3.0",
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "google-genai>=1.0.0",
    "pgvector>=0.3.0",
    "psycopg[binary]>=3.2.0",
    "psycopg2-binary>=2.9.0",
    "asyncpg>=0.30.0",
    "redis>=5.0.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "typer>=0.15.0",
    "rich>=13.0.0",
    "uvicorn>=0.34.0",
    "fastapi>=0.115.0",
    "python-dotenv>=1.0.0",
    "pyjwt>=2.8.0",
    "bcrypt>=4.0.0",
    "sqlalchemy>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[project.scripts]
completepay-agent = "src.cli.main:app"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true
exclude = ["tests/", "scripts/"]
